#327: Show player average in summary
