+++ CourageScores/ClientApp/src/components/division_fixtures/tournaments/superleague/MatchLog.js
@@ -1,7 +1,7 @@
 import {MatchLogRow} from "./MatchLogRow";
 import {MatchLogTableHeading} from "./MatchLogTableHeading";
-import {sum} from "../../../../helpers/collections";
 import {useApp} from "../../../../AppContainer";
+import {getPlayerOverallAverage} from "../../../../helpers/superleague";
 
 export function MatchLog({ tournamentData, fixture, saygDataMap }) {
     const { onError } = useApp();
@@ -29,20 +29,6 @@ export function MatchLog({ tournamentData, fixture, saygDataMap }) {
         }
 
         return throws;
-    }
-
-    function getPlayerOverallAverage(saygData, sideName) {
-        const metrics = Object.keys(saygData.legs).map(legIndex => {
-            const leg = saygData.legs[legIndex];
-            const side = leg[sideName];
-
-            return {
-                score: sum(side.throws, thr => thr.score),
-                noOfDarts: sum(side.throws, thr => thr.noOfDarts),
-            };
-        });
-
-        return sum(metrics, m => m.score) / sum(metrics, m => m.noOfDarts);
     }
 
     try {
